authz: move AllowAllPermissionPolicy to be private and hidden inside permission-backend
